Crate parquet_variant

Implementation of Variant Binary Encoding from Apache Parquet.

Main APIs

Variant: Represents a variant value, which can be an object, list, or primitive.

VariantBuilder: For building Variant values.

🚧 Work In Progress

This crate is under active development and is not yet ready for production use. If you are interested in helping, you can find more information on the GitHub Variant issue

Modules

builder 🔒

decoder 🔒

path 🔒

utils 🔒

variant 🔒

Structs

ListBuilder

A builder for creating Variant::List values.

ObjectBuilder

A builder for creating Variant::Object values.

ShortString

A Variant ShortString

VariantBuilder

Top level builder for Variant values

VariantDecimal4

Represents a 4-byte decimal value in the Variant format.

VariantDecimal8

Represents an 8-byte decimal value in the Variant format.

VariantDecimal16

Represents an 16-byte decimal value in the Variant format.

VariantList

Variant Array.

VariantMetadata

Variant Metadata

VariantObject

A Variant Object (struct with named fields).

VariantPath

Represents a qualified path to a potential subfield or index of a variant value.

Enums

Variant

Represents a Parquet Variant

VariantPathElement

Element of a VariantPath that can be a field name or an index.

Traits

VariantBuilderExt

Extends VariantBuilder to help building nested Variants