parquet_variant

Module builder

Re-exports

pub use list::*;

pub use metadata::*;

pub use object::*;

Modules

list 🔒

metadata 🔒

object 🔒

Macros

variant_append_value 🔒

Macro to generate the match statement for each append_variant, try_append_variant, and append_variant_bytes – they each have slightly different handling for object and list handling.

Structs

ParentState

Tracks information needed to correctly finalize a nested builder.

ValueBuilder

Wrapper around a Vec<u8> that provides methods for appending primitive values, variant types, and metadata.

VariantBuilder

Top level builder for Variant values

Constants

BASIC_TYPE_BITS 🔒

UNIX_EPOCH_DATE 🔒

Traits

BuilderSpecificState

A trait for managing state specific to different builder types.

VariantBuilderExt

Extends VariantBuilder to help building nested Variants

Functions

int_size 🔒

primitive_header 🔒

short_string_header 🔒

write_offset_at_pos 🔒

Write little-endian integer to buffer at a specific position