parquet_variant::builder

Enum ParentState

enum ParentState<'a> {
    Variant {
        buffer: &'a mut ValueBuffer,
        metadata_builder: &'a mut MetadataBuilder,
    },
    List {
        buffer: &'a mut ValueBuffer,
        metadata_builder: &'a mut MetadataBuilder,
        parent_value_offset_base: usize,
        offsets: &'a mut Vec<usize>,
    },
    Object {
        buffer: &'a mut ValueBuffer,
        metadata_builder: &'a mut MetadataBuilder,
        fields: &'a mut IndexMap<u32, usize>,
        field_name: &'a str,
        parent_value_offset_base: usize,
    },
}

Tracks information needed to correctly finalize a nested builder, for each parent builder type.

A child builder has no effect on its parent unless/until its finalize method is called, at which point the child appends the new value to the parent. As a (desirable) side effect, creating a parent state instance captures mutable references to a subset of the parent’s fields, rendering the parent object completely unusable until the parent state goes out of scope. This ensures that at most one child builder can exist at a time.

The redundancy in buffer and metadata_builder is because all the references come from the parent, and we cannot “split” a mutable reference across two objects (parent state and the child builder that uses it). So everything has to be here. Rust layout optimizations should treat the variants as a union, so that accessing a buffer or metadata_builder is branch-free.

Variants

Variant

Fields

buffer: &'a mut ValueBuffer

metadata_builder: &'a mut MetadataBuilder

List

Fields

buffer: &'a mut ValueBuffer

metadata_builder: &'a mut MetadataBuilder

parent_value_offset_base: usize

offsets: &'a mut Vec<usize>

Object

Fields

buffer: &'a mut ValueBuffer

metadata_builder: &'a mut MetadataBuilder

fields: &'a mut IndexMap<u32, usize>

field_name: &'a str

parent_value_offset_base: usize

Implementations

impl ParentState<'_>

fn buffer(&mut self) -> &mut ValueBuffer

fn metadata_builder(&mut self) -> &mut MetadataBuilder

fn finish(&mut self, starting_offset: usize)

fn buffer_and_metadata_builder( &mut self, ) -> (&mut ValueBuffer, &mut MetadataBuilder)

Return mutable references to the buffer and metadata builder that this parent state is using.

fn buffer_current_offset(&self) -> usize

fn metadata_current_offset(&self) -> usize