parquet_variant::builder

Module metadata

Structs

ReadOnlyMetadataBuilder

A metadata builder that cannot register new field names, and merely returns the field id associated with a known field name. This is useful for variant unshredding operations, where the metadata column is fixed and – per variant shredding spec – already contains all field names from the typed_value column. It is also useful when projecting a subset of fields from a variant object value, since the bytes can be copied across directly without re-encoding their field ids.

WritableMetadataBuilder

Builder for constructing metadata for Variant values.

Traits

MetadataBuilder

A trait for building variant metadata dictionaries, to be used in conjunction with a ValueBuilder. The trait provides methods for managing field names and their IDs, as well as rolling back a failed builder operation that might have created new field ids.

Functions

write_offset 🔒

Write little-endian integer to buffer