parquet_variant::builder::metadata

Struct WritableMetadataBuilder

pub struct WritableMetadataBuilder {
    pub(crate) field_names: IndexSet<String>,
    pub(crate) is_sorted: bool,
    metadata_buffer: Vec<u8>,
}

Builder for constructing metadata for Variant values.

This is used internally by the VariantBuilder to construct the metadata

You can use an existing Vec<u8> as the metadata buffer by using the from impl.

Fields
field_names: IndexSet<String>is_sorted: boolmetadata_buffer: Vec<u8>

Output buffer. Metadata is written to the end of this buffer

Implementations

impl WritableMetadataBuilder

pub fn upsert_field_name(&mut self, field_name: &str) -> u32

Upsert field name to dictionary, return its ID

pub fn offset(&self) -> usize

The current length of the underlying metadata buffer

fn num_field_names(&self) -> usize

Returns the number of field names stored in the metadata builder. Note: this method should be the only place to call self.field_names.len()

Panics

If the number of field names exceeds the maximum allowed value for u32.

fn field_name(&self, i: usize) -> &str

fn metadata_size(&self) -> usize

pub fn finish(&mut self) -> usize

Finalizes the metadata dictionary and appends its serialized bytes to the underlying buffer, returning the resulting Self::offset. The builder state is reset and ready to start building a new metadata dictionary.

pub fn into_inner(self) -> Vec<u8>

Returns the inner buffer, consuming self without finalizing any in progress metadata.