parquet_variant::builder

Struct ValueBuffer

struct ValueBuffer(Vec<u8>);

Wrapper around a Vec<u8> that provides methods for appending primitive values, variant types, and metadata.

This is used internally by the builders to construct the the value field for Variant values.

You can reuse an existing Vec<u8> by using the from impl

Tuple Fields
0: Vec<u8>
Implementations

impl ValueBuffer

fn new() -> Self

Construct a ValueBuffer that will write to a new underlying Vec

impl ValueBuffer

fn append_u8(&mut self, term: u8)

fn append_slice(&mut self, other: &[u8])

fn append_primitive_header(&mut self, primitive_type: VariantPrimitiveType)

fn into_inner(self) -> Vec<u8>

fn inner_mut(&mut self) -> &mut Vec<u8>

fn append_null(&mut self)

fn append_bool(&mut self, value: bool)

fn append_int8(&mut self, value: i8)

fn append_int16(&mut self, value: i16)

fn append_int32(&mut self, value: i32)

fn append_int64(&mut self, value: i64)

fn append_float(&mut self, value: f32)

fn append_double(&mut self, value: f64)

fn append_date(&mut self, value: NaiveDate)

fn append_timestamp_micros(&mut self, value: DateTime<Utc>)

fn append_timestamp_ntz_micros(&mut self, value: NaiveDateTime)

fn append_decimal4(&mut self, decimal4: VariantDecimal4)

fn append_decimal8(&mut self, decimal8: VariantDecimal8)

fn append_decimal16(&mut self, decimal16: VariantDecimal16)

fn append_binary(&mut self, value: &[u8])

fn append_short_string(&mut self, value: ShortString<'_>)

fn append_string(&mut self, value: &str)

fn append_object( &mut self, metadata_builder: &mut MetadataBuilder, obj: VariantObject<'_, '_>, )

fn try_append_object( &mut self, metadata_builder: &mut MetadataBuilder, obj: VariantObject<'_, '_>, ) -> Result<(), ArrowError>

fn append_list( &mut self, metadata_builder: &mut MetadataBuilder, list: VariantList<'_, '_>, )

fn try_append_list( &mut self, metadata_builder: &mut MetadataBuilder, list: VariantList<'_, '_>, ) -> Result<(), ArrowError>

fn offset(&self) -> usize

fn new_object<'a>( &'a mut self, metadata_builder: &'a mut MetadataBuilder, ) -> ObjectBuilder<'a>

fn new_list<'a>( &'a mut self, metadata_builder: &'a mut MetadataBuilder, ) -> ListBuilder<'a>

fn append_variant<'m, 'd>( &mut self, variant: Variant<'m, 'd>, metadata_builder: &mut MetadataBuilder, )

Appends a variant to the buffer.

Panics

This method will panic if the variant contains duplicate field names in objects when validation is enabled. For a fallible version, use ValueBuffer::try_append_variant

fn try_append_variant<'m, 'd>( &mut self, variant: Variant<'m, 'd>, metadata_builder: &mut MetadataBuilder, ) -> Result<(), ArrowError>

Appends a variant to the buffer

fn append_header_start_from_buf_pos( &mut self, start_pos: usize, header_byte: u8, is_large: bool, num_fields: usize, ) -> usize

Writes out the header byte for a variant object or list, from the starting position of the buffer, will return the position after this write

fn append_offset_array_start_from_buf_pos( &mut self, start_pos: usize, offsets: impl IntoIterator<Item = usize>, data_size: Option<usize>, nbytes: u8, ) -> usize

Writes out the offsets for an array of offsets, including the final offset (data size). from the starting position of the buffer, will return the position after this write