parquet_variant

Module decoder

Enums

OffsetSizeBytes 🔒

Used to unpack offset array entries such as metadata dictionary offsets or object/array value offsets. Also used to unpack object field ids. These are always derived from a two-bit XXX_size_minus_one field in the corresponding header byte.

VariantBasicType

The basic type of a Variant value, encoded in the first two bits of the header byte.

VariantPrimitiveType

The type of VariantBasicType::Primitive, for a primitive Variant value.

Functions

decode_binary 🔒

Decodes a Binary from the value section of a variant.

decode_date 🔒

Decodes a Date from the value section of a variant.

decode_decimal4 🔒

Decodes a Decimal4 from the value section of a variant.

decode_decimal8 🔒

Decodes a Decimal8 from the value section of a variant.

decode_decimal16 🔒

Decodes a Decimal16 from the value section of a variant.

decode_double 🔒

Decodes a Double from the value section of a variant.

decode_float 🔒

Decodes a Float from the value section of a variant.

decode_int8 🔒

Decodes an Int8 from the value section of a variant.

decode_int16 🔒

Decodes an Int16 from the value section of a variant.

decode_int32 🔒

Decodes an Int32 from the value section of a variant.

decode_int64 🔒

Decodes an Int64 from the value section of a variant.

decode_long_string 🔒

Decodes a long string from the value section of a variant.

decode_short_string 🔒

Decodes a short string from the value section of a variant.

decode_time_ntz 🔒

decode_timestamp_micros 🔒

Decodes a TimestampMicros from the value section of a variant.

decode_timestamp_nanos 🔒

Decodes a TimestampNanos from the value section of a variant.

decode_timestampntz_micros 🔒

Decodes a TimestampNtzMicros from the value section of a variant.

decode_timestampntz_nanos 🔒

Decodes a TimestampNtzNanos from the value section of a variant.

decode_uuid 🔒

Decodes a UUID from the value section of a variant.

get_basic_type 🔒

Extracts the basic type from a header byte

get_primitive_type 🔒

Extract the primitive type from a Variant value-metadata byte

map_bytes_to_offsets 🔒

Converts a byte buffer to offset values based on the specific offset size