parquet_variant

Struct ReadOnlyMetadataBuilder

pub struct ReadOnlyMetadataBuilder<'m> {
    metadata: &'m VariantMetadata<'m>,
    known_field_names: HashMap<&'m str, u32>,
}

A metadata builder that cannot register new field names, and merely returns the field id associated with a known field name. This is useful for variant unshredding operations, where the metadata column is fixed and – per variant shredding spec – already contains all field names from the typed_value column. It is also useful when projecting a subset of fields from a variant object value, since the bytes can be copied across directly without re-encoding their field ids.

NOTE: Self::finish is a no-op. If the intent is to make a copy of the underlying bytes each time finish is called, a different trait impl will be needed.

Fields
metadata: &'m VariantMetadata<'m>known_field_names: HashMap<&'m str, u32>
Implementations

impl<'m> ReadOnlyMetadataBuilder<'m>

pub fn new(metadata: &'m VariantMetadata<'m>) -> Self

Creates a new read-only metadata builder from the given metadata dictionary.