parquet_variant

Struct VariantBuilder

pub struct VariantBuilder {
    buffer: Vec<u8>,
    dict: HashMap<String, u32>,
    dict_keys: Vec<String>,
}

Builder for Variant values

Example: create a Primitive Int8

let mut builder = VariantBuilder::new();
builder.append_value(Variant::Int8(42));
// Finish the builder to get the metadata and value
let (metadata, value) = builder.finish();
// use the Variant API to verify the result
let variant = Variant::try_new(&metadata, &value).unwrap();
assert_eq!(variant, Variant::Int8(42));

Example: Create a Variant::Object

This example shows how to create an object with two fields:

{
 "first_name": "Jiaying",
 "last_name": "Li"
}

let mut builder = VariantBuilder::new();
// Create an object builder that will write fields to the object
let mut object_builder = builder.new_object();
object_builder.append_value("first_name", "Jiaying");
object_builder.append_value("last_name", "Li");
object_builder.finish();
// Finish the builder to get the metadata and value
let (metadata, value) = builder.finish();
// use the Variant API to verify the result
let variant = Variant::try_new(&metadata, &value).unwrap();
let Variant::Object(variant_object) = variant else {
  panic!("unexpected variant type")
};
assert_eq!(
  variant_object.field_by_name("first_name").unwrap(),
  Some(Variant::from("Jiaying"))
);
assert_eq!(
  variant_object.field_by_name("last_name").unwrap(),
  Some(Variant::from("Li"))
);

Example: Create a Variant::List (an Array)

This example shows how to create an array of integers: [1, 2, 3].

 let mut builder = VariantBuilder::new();
 // Create a builder that will write elements to the list
 let mut list_builder = builder.new_list();
 list_builder.append_value(1i8);
 list_builder.append_value(2i8);
 list_builder.append_value(3i8);
 list_builder.finish();
// Finish the builder to get the metadata and value
let (metadata, value) = builder.finish();
// use the Variant API to verify the result
let variant = Variant::try_new(&metadata, &value).unwrap();
let Variant::List(variant_list) = variant else {
  panic!("unexpected variant type")
};
// Verify the list contents
assert_eq!(variant_list.get(0).unwrap(), Variant::Int8(1));
assert_eq!(variant_list.get(1).unwrap(), Variant::Int8(2));
assert_eq!(variant_list.get(2).unwrap(), Variant::Int8(3));

Example: Variant::List of Variant::Objects

THis example shows how to create an list of objects:

[
 {
  "first_name": "Jiaying",
 "last_name": "Li"
},
  {
   "first_name": "Malthe",
   "last_name": "Karbo"
}
]

TODO

Fields
buffer: Vec<u8>dict: HashMap<String, u32>dict_keys: Vec<String>
Implementations

impl VariantBuilder

pub fn new() -> Self

fn append_null(&mut self)

fn append_bool(&mut self, value: bool)

fn append_int8(&mut self, value: i8)

fn append_int16(&mut self, value: i16)

fn append_int32(&mut self, value: i32)

fn append_int64(&mut self, value: i64)

fn append_float(&mut self, value: f32)

fn append_double(&mut self, value: f64)

fn append_date(&mut self, value: NaiveDate)

fn append_timestamp_micros(&mut self, value: DateTime<Utc>)

fn append_timestamp_ntz_micros(&mut self, value: NaiveDateTime)

fn append_decimal4(&mut self, integer: i32, scale: u8)

fn append_decimal8(&mut self, integer: i64, scale: u8)

fn append_decimal16(&mut self, integer: i128, scale: u8)

fn append_binary(&mut self, value: &[u8])

fn append_short_string(&mut self, value: ShortString<'_>)

fn append_string(&mut self, value: &str)

fn add_key(&mut self, key: &str) -> u32

Add key to dictionary, return its ID

fn offset(&self) -> usize

pub fn new_list(&mut self) -> ListBuilder<'_>

Create an ListBuilder for creating Variant::List values.

See the examples on VariantBuilder for usage.

pub fn new_object(&mut self) -> ObjectBuilder<'_>

Create an ObjectBuilder for creating Variant::Object values.

See the examples on VariantBuilder for usage.

pub fn finish(self) -> (Vec<u8>, Vec<u8>)

pub fn append_value<'m, 'd, T: Into<Variant<'m, 'd>>>(&mut self, value: T)