parquet_variant

Module utils

Functions

array_from_slice 🔒

first_byte_from_slice 🔒

overflow_error 🔒

Helper for reporting integer overflow errors in a consistent way.

slice_from_slice 🔒

slice_from_slice_at_offset 🔒

Helper to safely slice bytes with offset calculations.

string_from_slice 🔒

Helper to get a &str from a slice at the given offset and range, or an error if invalid.

try_binary_search_range_by 🔒

Performs a binary search over a range using a fallible key extraction function; a failed key extraction immediately terminats the search.

validate_fallible_iterator 🔒

Attempts to prove a fallible iterator is actually infallible in practice, by consuming every element and returning the first error (if any).