Crate parquet_variant

Implementation of Variant Binary Encoding from Apache Parquet.

🚧 Work In Progress

This crate is under active development and is not yet ready for production use. If you are interested in helping, you can find more information on the GitHub Variant issue

Modules

builder 🔒

decoder 🔒

from_json 🔒

Module for parsing JSON strings as Variant

to_json 🔒

Module for converting Variant data to JSON format

utils 🔒

variant 🔒

Structs

ListBuilder

A builder for creating Variant::List values.

ObjectBuilder

A builder for creating Variant::Object values.

ShortString

A Variant ShortString

VariantBuilder

Top level builder for Variant values

VariantDecimal4

Represents a 4-byte decimal value in the Variant format.

VariantDecimal8

Represents an 8-byte decimal value in the Variant format.

VariantDecimal16

Represents an 16-byte decimal value in the Variant format.

VariantList

Variant Array.

VariantMetadata

Variant Metadata

VariantObject

A Variant Object (struct with named fields).

Enums

Variant

Represents a Parquet Variant

Functions

json_to_variant

Converts a JSON string to Variant using VariantBuilder. The resulting value and metadata buffers can be extracted using builder.finish()

variant_to_json

This function writes JSON directly to any type that implements Write, making it efficient for streaming or when you want to control the output destination.

variant_to_json_string

Convert Variant to JSON String

variant_to_json_value

Convert Variant to serde_json::Value