parquet_variant

Module builder

Macros

variant_append_value 🔒

Macro to generate the match statement for each append_variant, try_append_variant, and append_variant_bytes – they each have slightly different handling for object and list handling.

Structs

ListBuilder

A builder for creating Variant::List values.

ListState

Internal state for list building

ObjectBuilder

A builder for creating Variant::Object values.

ObjectFieldBuilder

A VariantBuilderExt that inserts a new field into a variant object.

ObjectState

Internal state for object building

ParentState

Tracks information needed to correctly finalize a nested builder.

ReadOnlyMetadataBuilder

A metadata builder that cannot register new field names, and merely returns the field id associated with a known field name. This is useful for variant unshredding operations, where the metadata column is fixed and – per variant shredding spec – already contains all field names from the typed_value column. It is also useful when projecting a subset of fields from a variant object value, since the bytes can be copied across directly without re-encoding their field ids.

ValueBuilder

Wrapper around a Vec<u8> that provides methods for appending primitive values, variant types, and metadata.

VariantBuilder

Top level builder for Variant values

WritableMetadataBuilder

Builder for constructing metadata for Variant values.

Constants

BASIC_TYPE_BITS 🔒

UNIX_EPOCH_DATE 🔒

Traits

BuilderSpecificState

A trait for managing state specific to different builder types.

MetadataBuilder

A trait for building variant metadata dictionaries, to be used in conjunction with a ValueBuilder. The trait provides methods for managing field names and their IDs, as well as rolling back a failed builder operation that might have created new field ids.

VariantBuilderExt

Extends VariantBuilder to help building nested Variants

Functions

append_packed_u32 🔒

Append value_size bytes of given value into dest.

array_header 🔒

int_size 🔒

object_header 🔒

primitive_header 🔒

short_string_header 🔒

write_offset 🔒

Write little-endian integer to buffer

write_offset_at_pos 🔒

Write little-endian integer to buffer at a specific position