parquet_variant

Module builder

Structs

ListBuilder

A builder for creating Variant::List values.

MetadataBuilder ğŸ”’

Builder for constructing metadata for Variant values.

ObjectBuilder

A builder for creating Variant::Object values.

ValueBuffer ğŸ”’

Wrapper around a Vec<u8> that provides methods for appending primitive values, variant types, and metadata.

VariantBuilder

Top level builder for Variant values

Enums

ParentState ğŸ”’

Tracks information needed to correctly finalize a nested builder, for each parent builder type.

Constants

BASIC_TYPE_BITS ğŸ”’

UNIX_EPOCH_DATE ğŸ”’

Traits

VariantBuilderExt

Extends VariantBuilder to help building nested Variants

Functions

append_packed_u32 ğŸ”’

Append value_size bytes of given value into dest.

array_header ğŸ”’

int_size ğŸ”’

object_header ğŸ”’

primitive_header ğŸ”’

short_string_header ğŸ”’

write_offset ğŸ”’

Write little-endian integer to buffer

write_offset_at_pos ğŸ”’

Write little-endian integer to buffer at a specific position