parquet_variant

Struct ObjectBuilder

pub struct ObjectBuilder<'a> {
    parent_state: ParentState<'a>,
    fields: IndexMap<u32, usize>,
    parent_value_offset_base: usize,
    parent_metadata_offset_base: usize,
    has_been_finished: bool,
    validate_unique_fields: bool,
    duplicate_fields: HashSet<u32>,
}

A builder for creating Variant::Object values.

See the examples on VariantBuilder for usage.

Fields
parent_state: ParentState<'a>fields: IndexMap<u32, usize>parent_value_offset_base: usize

The starting offset in the parent’s buffer where this object starts

parent_metadata_offset_base: usize

The starting offset in the parent’s metadata buffer where this object starts used to truncate the written fields in drop if the current object has not been finished

has_been_finished: bool

Whether the object has been finished, the written content of the current object will be truncated in drop if has_been_finished is false

validate_unique_fields: boolduplicate_fields: HashSet<u32>

Set of duplicate fields to report for errors

Implementations

impl<'a> ObjectBuilder<'a>

fn new(parent_state: ParentState<'a>, validate_unique_fields: bool) -> Self

pub fn insert<'m, 'd, T: Into<Variant<'m, 'd>>>(&mut self, key: &str, value: T)

Add a field with key and value to the object

See Also

ObjectBuilder::try_insert for a fallible version.

ObjectBuilder::with_field for a builder-style API.

Panics

This method will panic if the variant contains duplicate field names in objects when validation is enabled. For a fallible version, use ObjectBuilder::try_insert

pub fn try_insert<'m, 'd, T: Into<Variant<'m, 'd>>>( &mut self, key: &str, value: T, ) -> Result<(), ArrowError>

Add a field with key and value to the object

See Also

ObjectBuilder::insert for a infallabel version

ObjectBuilder::try_with_field for a builder-style API.

Note

When inserting duplicate keys, the new value overwrites the previous mapping, but the old value remains in the buffer, resulting in a larger variant

pub fn with_field<'m, 'd, T: Into<Variant<'m, 'd>>>( self, key: &str, value: T, ) -> Self

Builder style API for adding a field with key and value to the object

Same as ObjectBuilder::insert, but returns self for chaining.

pub fn try_with_field<'m, 'd, T: Into<Variant<'m, 'd>>>( self, key: &str, value: T, ) -> Result<Self, ArrowError>

Builder style API for adding a field with key and value to the object

Same as ObjectBuilder::try_insert, but returns self for chaining.

pub fn with_validate_unique_fields(self, validate_unique_fields: bool) -> Self

Enables validation for unique field keys when inserting into this object.

When this is enabled, calling ObjectBuilder::finish will return an error if any duplicate field keys were added using ObjectBuilder::insert.

fn parent_state<'b>(&'b mut self, key: &'b str) -> (ParentState<'b>, bool)

pub fn new_object<'b>(&'b mut self, key: &'b str) -> ObjectBuilder<'b>

Returns an object builder that can be used to append a new (nested) object to this object.

WARNING: The builder will have no effect unless/until ObjectBuilder::finish is called.

pub fn new_list<'b>(&'b mut self, key: &'b str) -> ListBuilder<'b>

Returns a list builder that can be used to append a new (nested) list to this object.

WARNING: The builder will have no effect unless/until ListBuilder::finish is called.

pub fn finish(self) -> Result<(), ArrowError>

Finalizes this object and appends it to its parent, which otherwise remains unmodified.