parquet_variant

Struct VariantDecimal16

pub struct VariantDecimal16 {
    integer: i128,
    scale: u8,
}

Represents an 16-byte decimal value in the Variant format.

This struct stores a decimal number using a 128-bit signed integer for the coefficient and an 8-bit unsigned integer for the scale (number of decimal places). Its precision is between 19 and 38 digits.

For valid precision and scale values, see the Variant specification:

https://github.com/apache/parquet-format/blob/87f2c8bf77eefb4c43d0ebaeea1778bd28ac3609/VariantEncoding.md?plain=1#L418-L420

Example: Create a VariantDecimal16

// Create a value representing the decimal 12345678901234567.890
let decimal = VariantDecimal16::try_new(12345678901234567890, 3).unwrap();

Fields
integer: i128scale: u8
Implementations

impl VariantDecimal16

const MAX_PRECISION: u8 = 38u8

const MAX_UNSCALED_VALUE: u128 = 99_999_999_999_999_999_999_999_999_999_999_999_999u128

pub fn try_new(integer: i128, scale: u8) -> Result<Self, ArrowError>

pub fn integer(&self) -> i128

Returns the underlying value of the decimal.

For example, if the decimal is 12345678901234567.890, this will return 12345678901234567890.

pub fn scale(&self) -> u8

Returns the scale of the decimal (how many digits after the decimal point).

For example, if the decimal is 12345678901234567.890, this will return 3.